// src/common/pb/todo.pb.ts
export interface GetTodoListRequest {
  accountID: number;
}

export interface GetTodoListResponse {
  todo: Todo[];
}

export interface Todo {
  todoID: number;
  todoName: string;
  status: number;
  createdAt: Timestamp | undefined;
  updatedAt: Timestamp | undefined;
}

export interface CreateTodoRequest {
  accountID: number;
  name: string;
}

export interface CreateTodoResponse {
  todoID: number;
  todoName: string;
  status: number;
  createdAt: Timestamp | undefined;
  updatedAt: Timestamp | undefined;
}

export interface GetTodoRequest {
  accountID: number;
  todoID: number;
}

export interface GetTodoResponse {
  todoID: number;
  todoName: string;
  status: number;
  createdAt: Timestamp | undefined;
  updatedAt: Timestamp | undefined;
}

export interface DeleteTodoRequest {
  accountID: number;
  todoID: number;
}

export interface DeleteTodoResponse {
  todoID: number;
  todoName: string;
  status: number;
  createdAt: Timestamp | undefined;
  updatedAt: Timestamp | undefined;
}

export interface CreateTaskRequest {
  todoID: number;
  accountID: number;
  name: string;
}

export interface CreateTaskResponse {
  taskID: number;
  taskName: string;
  status: number;
  createdAt: Timestamp | undefined;
  updatedAt: Timestamp | undefined;
}

export interface GetTaskListRequest {
  todoID: number;
  accountID: number;
}

export interface GetTaskListResponse {
  tasks: Task[];
}

export interface Task {
  taskID: number;
  taskName: string;
  status: number;
  createdAt: Timestamp | undefined;
  updatedAt: Timestamp | undefined;
}

export interface UpdateTaskRequest {
  todoID: number;
  accountID: number;
  taskID: number;
  status: number;
}

export interface UpdateTaskResponse {
  taskID: number;
  taskName: string;
  status: number;
  createdAt: Timestamp | undefined;
  updatedAt: Timestamp | undefined;
}

export interface DeleteTaskRequest {
  todoID: number;
  accountID: number;
  taskID: number;
}

export interface DeleteTaskResponse {
  taskID: number;
  taskName: string;
  status: number;
  createdAt: Timestamp | undefined;
  updatedAt: Timestamp | undefined;
}

export const BTODO_TODO_PACKAGE_NAME = "btodo.todo";

export interface TodoServiceClient {
  getTodoList(request: GetTodoListRequest): Observable<GetTodoListResponse>;

  createTodo(request: CreateTodoRequest): Observable<CreateTodoResponse>;

  getTodo(request: GetTodoRequest): Observable<GetTodoResponse>;

  deleteTodo(request: DeleteTodoRequest): Observable<DeleteTodoResponse>;

  createTask(request: CreateTaskRequest): Observable<CreateTaskResponse>;

  getTaskList(request: GetTaskListRequest): Observable<GetTaskListResponse>;

  updateTask(request: UpdateTaskRequest): Observable<UpdateTaskResponse>;

  deleteTask(request: DeleteTaskRequest): Observable<DeleteTaskResponse>;
}
// ...
export const TODO_SERVICE_NAME = "TodoService";